refactor: split ui.py into two files
