+++ ui/__init__.py
@@ -1,0 +1,7 @@
+from .completion import ViewCompletionManager
+from .panel_completion import ViewPanelCompletionManager
+
+__all__ = (
+    "ViewCompletionManager",
+    "ViewPanelCompletionManager",
+)
